+++ src/Plugins/Module/Location/src/LocationWidget.h
@@ -98,7 +98,8 @@ private slots:
     void onSearchTextChanged() noexcept;
     void searchText() noexcept;
     void onTypeOptionToggled(const QVariant &optionValue, bool enable) noexcept;
-    void resetSearchFilter() noexcept;
+    void resetFilters() noexcept;
+    void resetDefaultValues() noexcept;
 
     void onCellSelected(int row, int column) noexcept;
     void onCellChanged(int row, int column) noexcept;